requirements.txt
Flask==3.1.1
supabase==2.4.0
requests==2.31.0
opencv-contrib-python==4.11.0.86
numpy==2.2.4
Pillow==11.2.1
